Fix components path
- Commented PropTypes
