+++ src/components/error/error.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 
 export default function ErrorComponent() {
   return (
