+++ src/components/favorite/favorite-footer/favorite-footer.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 
 export default function FavoriteFooter() {
   return (
