+++ src/redux/offers-favorites/offers-favorites.js
@@ -1,6 +1,6 @@
 import {getFavorites, setFavorite} from "../../api/clients";
-import {parseHotel} from "../mapping/hotel-parser.js";
-import {parseHotels} from "../mapping/hotels-pareser.js";
+import {parseHotel} from "../mapping/hotel-parser";
+import {parseHotels} from "../mapping/hotels-pareser";
 import {produce} from 'immer';
 
 const ActionType = {
